refactor and added tittle for tasks sorteb by status
