+++ app/(dashboard)/[id]/l/[projectId]/page.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import React from "react";
-import ViewsBarContainer from "@/app/(dashboard)/_components/ViewsBarContainer";
-import WidgetHeader from "@/app/(dashboard)/_components/WidgetHeader";
+import ViewsBarContainer from "@/app/(dashboard)/_components/SubNavBar/ViewsBarContainer";
+import WidgetHeader from "@/app/(dashboard)/_components/SubNavBar/WidgetHeader";
 import ButtonVariant2 from "@/components/ButtonVariant2";
 import TasksList from "./components/TasksList";
 
@@ -11,13 +11,15 @@ export default function Page() {
     <div className="flex flex-col gap-4 w-full h-full">
       <div>
         <WidgetHeader>
-          <ButtonVariant2 isActive={true} className={`items-center h-8 font-semibold `}>
+          <ButtonVariant2
+            isActive={true}
+            className={`items-center h-8 font-semibold `}
+          >
             Everything
           </ButtonVariant2>
         </WidgetHeader>
         <ViewsBarContainer />
       </div>
-      {/* Example for using the Table with filtered tasks */}
       <TasksList />
     </div>
   );
